taint/test/mochitest/test_websocket.html

<!DOCTYPE HTML>
<html>
  <head>
    <meta charset="utf-8">
    <title>Check Location Sources</title>
    <script src="/tests/SimpleTest/SimpleTest.js"></script>
    <link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
    <script>

      function check_tainted(str) {
          ok(str.taint.length > 0, "Check tainted");
      }

      function check_untainted(str) {
          ok(str.taint.length == 0, "Check untainted");
      }

      SimpleTest.waitForExplicitFinish();
      function startTest() {
        let ws = new WebSocket("ws://mochi.test:8888/tests/taint/test/mochitest/file_websocket");

        ws.addEventListener("message", (event) => {
            check_tainted(event.data);
            SimpleTest.finish();
            ws.close();
        });

        ws.addEventListener("onload", (event) => {
            ws.send("untainted");
        });
      }


    </script>
  </head>
  <body onload="startTest()">
  </body>
</html>
